+++ src/main.cpp
@@ -13,6 +13,7 @@ static std::ostream* Debug = &std::cerr;
 
 extern const uint8_t mrbc_rj[];
 extern const uint8_t mrbc_ruby_formatter[];
+extern const uint8_t mrbc_method_accessor[];
 
 //! \brief Parses and represents command line arguments.
 struct ProgramOptions
@@ -193,6 +194,7 @@ int main(int argc, char** argv)
         rj_class->ivar_set("@debug", mrb_bool_value(opts.debug));
         rj_class->ivar_set("@output_format", opts.output_format);
         rb.load_irep(mrbc_ruby_formatter);
+        rb.load_irep(mrbc_method_accessor);
 
         *Debug << "reading from stdin" << std::endl;
         if (!rb.eval("item = JSON.parse(STDIN.read)"))